requirements.txt
django
psycopg2-binary
pillow
djangorestframework
python-decouple
django-phonenumber-field
phonenumbers
twilio
djangorestframework-simplejwt
drf-yasg